make code origin a stage
