+++ packages/dd-trace/src/plugins/stages/code-origin.js
@@ -1,0 +1,15 @@
+'use strict'
+
+const { exitTags } = require('../../../../datadog-code-origin')
+
+const addExitCodeOrigin = Symbol('integration.pipeline.add_exit_code_origin')
+
+const exitCodeOrigin = {
+  name: 'exit-code-origin',
+  requires: ['tracing'],
+  start (frame) {
+    frame[addExitCodeOrigin](exitTags, exitCodeOrigin.start)
+  },
+}
+
+module.exports = { addExitCodeOrigin, exitCodeOrigin }
